requirements.txt
coverage==4.5.1
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.8.17
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
